+++ ChannelIO/Source/Controllers/BaseViewController.swift
@@ -7,7 +7,6 @@
 //
 
 import UIKit
-import UINavigationItem_Margin
 
 class BaseViewController: UIViewController {
 
@@ -27,8 +26,6 @@ class BaseViewController: UIViewController {
 
   override func viewDidLoad() {
     super.viewDidLoad()
-    self.navigationItem.leftMargin = 0
-    self.navigationItem.rightMargin = 0
     self.view.setNeedsUpdateConstraints()
   }